First steps towards support for deeply nested resources.
When building the prefix for a nested resource, HyRes now looks to see if the
'nest' resource is itself nested, and if so adds the necessary prefix extensions
as far back as necessary.  To enable deep nesting, simple add the according
"belongs_to :nest_model, :nested => :true" line to each model as appropriate.